Code review changes 3.
