+++ hive-metastore/src/main/java/org/apache/iceberg/hive/ClientPool.java
@@ -19,130 +19,10 @@
 
 package org.apache.iceberg.hive;
 
-import java.io.Closeable;
-import java.util.ArrayDeque;
-import java.util.Deque;
-import org.apache.iceberg.relocated.com.google.common.annotations.VisibleForTesting;
-import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public abstract class ClientPool<C, E extends Exception> implements Closeable {
-  private static final Logger LOG = LoggerFactory.getLogger(ClientPool.class);
-
-  private final int poolSize;
-  private final Deque<C> clients;
-  private final Class<? extends E> reconnectExc;
-  private final Object signal = new Object();
-  private volatile int currentSize;
-  private boolean closed;
-
-  ClientPool(int poolSize, Class<? extends E> reconnectExc) {
-    this.poolSize = poolSize;
-    this.reconnectExc = reconnectExc;
-    this.clients = new ArrayDeque<>(poolSize);
-    this.currentSize = 0;
-    this.closed = false;
-  }
-
-  public interface Action<R, C, E extends Exception> {
+public interface ClientPool<C, E extends Exception> {
+  interface Action<R, C, E extends Exception> {
     R run(C client) throws E;
   }
 
-  public <R> R run(Action<R, C, E> action) throws E, InterruptedException {
-    C client = get();
-    try {
-      return action.run(client);
-
-    } catch (Exception exc) {
-      if (isConnectionException(exc)) {
-        try {
-          client = reconnect(client);
-        } catch (Exception ignored) {
-          // if reconnection throws any exception, rethrow the original failure
-          throw reconnectExc.cast(exc);
-        }
-
-        return action.run(client);
-      }
-
-      throw exc;
-
-    } finally {
-      release(client);
-    }
-  }
-
-  protected abstract C newClient();
-
-  protected abstract C reconnect(C client);
-
-  protected boolean isConnectionException(Exception exc) {
-    return reconnectExc.isInstance(exc);
-  }
-
-  protected abstract void close(C client);
-
-  @Override
-  public void close() {
-    this.closed = true;
-    try {
-      while (currentSize > 0) {
-        if (!clients.isEmpty()) {
-          synchronized (this) {
-            if (!clients.isEmpty()) {
-              C client = clients.removeFirst();
-              close(client);
-              currentSize -= 1;
-            }
-          }
-        }
-        if (clients.isEmpty() && currentSize > 0) {
-          // wake every second in case this missed the signal
-          synchronized (signal) {
-            signal.wait(1000);
-          }
-        }
-      }
-
-    } catch (InterruptedException e) {
-      Thread.currentThread().interrupt();
-      LOG.warn("Interrupted while shutting down pool. Some clients may not be closed.", e);
-    }
-  }
-
-  @VisibleForTesting
-  int poolSize() {
-    return poolSize;
-  }
-
-  private C get() throws InterruptedException {
-    Preconditions.checkState(!closed, "Cannot get a client from a closed pool");
-    while (true) {
-      if (!clients.isEmpty() || currentSize < poolSize) {
-        synchronized (this) {
-          if (!clients.isEmpty()) {
-            return clients.removeFirst();
-          } else if (currentSize < poolSize) {
-            C client = newClient();
-            currentSize += 1;
-            return client;
-          }
-        }
-      }
-      synchronized (signal) {
-        // wake every second in case this missed the signal
-        signal.wait(1000);
-      }
-    }
-  }
-
-  private void release(C client) {
-    synchronized (this) {
-      clients.addFirst(client);
-    }
-    synchronized (signal) {
-      signal.notify();
-    }
-  }
+  <R> R run(Action<R, C, E> action) throws E, InterruptedException;
 }
